wm: rework to use mmap'd files instead of sharing the framebuffer
- wm: move majority of drawing to wm
- kernel: fix mmap syscall, implement mmap() for tmpfs nodes

+++ userspace/libraries/libc/src/include/sys/gfx.h
@@ -2,8 +2,8 @@
 
 struct fbdev_bitblit {
     int target_buffer;
-    unsigned long *source;
-    unsigned long x, y, width, height;
+    unsigned int *source;
+    unsigned int x, y, width, height;
     int type;
 } __attribute__((packed));
